+++ graal-js/src/com.oracle.js.parser/src/com/oracle/js/parser/Parser.java
@@ -105,6 +105,7 @@ import static com.oracle.js.parser.TokenType.THIS;
 import static com.oracle.js.parser.TokenType.VAR;
 import static com.oracle.js.parser.TokenType.VOID;
 import static com.oracle.js.parser.TokenType.WHILE;
+import static com.oracle.js.parser.TokenType.WITH;
 import static com.oracle.js.parser.TokenType.YIELD;
 import static com.oracle.js.parser.TokenType.YIELD_STAR;
 
@@ -277,7 +278,7 @@ public class Parser extends AbstractParser {
     private static final String MSG_ASYNC_CONSTRUCTOR = "async.constructor";
     private static final String MSG_CONSTRUCTOR_FIELD = "constructor.field";
     private static final String MSG_DUPLICATE_DEFAULT_IN_SWITCH = "duplicate.default.in.switch";
-    private static final String MSG_DUPLICATE_IMPORT_ASSERTION = "duplicate.import.assertion";
+    private static final String MSG_DUPLICATE_IMPORT_ATTRIBUTE = "duplicate.import.attribute";
     private static final String MSG_DUPLICATE_LABEL = "duplicate.label";
     private static final String MSG_ESCAPED_KEYWORD = "escaped.keyword";
     private static final String MSG_EXPECTED_ARROW_PARAMETER = "expected.arrow.parameter";
@@ -5137,7 +5138,7 @@ public class Parser extends AbstractParser {
             next();
             List<Expression> arguments = new ArrayList<>();
             arguments.add(assignmentExpression(true, yield, await));
-            if (env.importAssertions && type == COMMARIGHT) {
+            if (env.importAttributes && type == COMMARIGHT) {
                 next();
                 if (type != RPAREN) {
                     arguments.add(assignmentExpression(true, yield, await));
@@ -7006,11 +7007,11 @@ public class Parser extends AbstractParser {
             long specifierToken = token;
             next();
             LiteralNode<TruffleString> specifier = LiteralNode.newInstance(specifierToken, moduleSpecifier);
-            Map<TruffleString, TruffleString> assertions = Map.of();
-            if (env.importAssertions && type == ASSERT && last != EOL) {
-                assertions = assertClause();
+            Map<TruffleString, TruffleString> attributes = Map.of();
+            if (env.importAttributes && ((type == WITH) || (type == ASSERT && last != EOL))) {
+                attributes = withClause();
             }
-            module.addModuleRequest(ModuleRequest.create(moduleSpecifier, assertions));
+            module.addModuleRequest(ModuleRequest.create(moduleSpecifier, attributes));
             module.addImport(new ImportNode(importToken, Token.descPosition(importToken), finish, specifier));
         } else {
             // import ImportClause FromClause ;
@@ -7053,13 +7054,13 @@ public class Parser extends AbstractParser {
             }
 
             FromNode fromNode = fromClause();
-            Map<TruffleString, TruffleString> assertions = Map.of();
-            if (env.importAssertions && type == ASSERT && last != EOL) {
-                assertions = assertClause();
+            Map<TruffleString, TruffleString> attributes = Map.of();
+            if (env.importAttributes && ((type == WITH) || (type == ASSERT && last != EOL))) {
+                attributes = withClause();
             }
             module.addImport(new ImportNode(importToken, Token.descPosition(importToken), finish, importClause, fromNode));
             TruffleString moduleSpecifier = fromNode.getModuleSpecifier().getValue();
-            ModuleRequest moduleRequest = ModuleRequest.create(moduleSpecifier, assertions);
+            ModuleRequest moduleRequest = ModuleRequest.create(moduleSpecifier, attributes);
             module.addModuleRequest(moduleRequest);
             for (int i = 0; i < importEntries.size(); i++) {
                 module.addImportEntry(importEntries.get(i).withFrom(moduleRequest));
@@ -7072,42 +7073,42 @@ public class Parser extends AbstractParser {
      * Parse assert clause.
      *
      * <pre>
-     *     assert { }
-     *     assert { AssertEntries ,opt }
+     *     AttributesKeyword { }
+     *     AttributesKeyword { WithEntries ,opt }
      * </pre>
      */
-    private Map<TruffleString, TruffleString> assertClause() {
-        assert type == ASSERT;
+    private Map<TruffleString, TruffleString> withClause() {
+        assert (type == ASSERT || type == WITH);
         next();
         expect(LBRACE);
-        Map<TruffleString, TruffleString> assertions = assertEntries();
+        Map<TruffleString, TruffleString> entries = withEntries();
         expect(RBRACE);
-        return assertions;
+        return entries;
     }
 
     /**
      * Parse assert entry.
      *
      * <pre>
-     * AssertEntries:
-     *     AssertionKey : StringLiteral
-     *     AssertionKey : StringLiteral , AssertEntries
-     * AssertKey:
+     * WithEntries:
+     *     AttributeKey : StringLiteral
+     *     AttributeKey : StringLiteral , WithEntries
+     * AttributeKey:
      *     IdentifierName
      *     StringLiteral
      * </pre>
      */
-    private Map<TruffleString, TruffleString> assertEntries() {
-        Map<TruffleString, TruffleString> assertions = new LinkedHashMap<>();
+    private Map<TruffleString, TruffleString> withEntries() {
+        Map<TruffleString, TruffleString> entries = new LinkedHashMap<>();
 
         while (type != RBRACE) {
             final long errorToken = token;
-            TruffleString assertionKey;
+            TruffleString attributeKey;
             if (type == STRING || type == ESCSTRING) {
-                assertionKey = (TruffleString) getValue();
+                attributeKey = (TruffleString) getValue();
                 next();
             } else {
-                assertionKey = getIdentifierName().getNameTS();
+                attributeKey = getIdentifierName().getNameTS();
             }
             expect(COLON);
             TruffleString value = null;
@@ -7117,10 +7118,10 @@ public class Parser extends AbstractParser {
             } else {
                 expect(STRING);
             }
-            if (assertions.containsKey(assertionKey)) {
-                throw error(AbstractParser.message(MSG_DUPLICATE_IMPORT_ASSERTION, assertionKey.toJavaStringUncached()), errorToken);
+            if (entries.containsKey(attributeKey)) {
+                throw error(AbstractParser.message(MSG_DUPLICATE_IMPORT_ATTRIBUTE, attributeKey.toJavaStringUncached()), errorToken);
             } else {
-                assertions.put(assertionKey, value);
+                entries.put(attributeKey, value);
             }
             if (type == COMMARIGHT) {
                 next();
@@ -7128,7 +7129,7 @@ public class Parser extends AbstractParser {
                 break;
             }
         }
-        return assertions;
+        return entries;
     }
 
     /**
@@ -7241,7 +7242,7 @@ public class Parser extends AbstractParser {
      */
     private void exportDeclaration(ParserContextModuleNode module) {
         final long exportToken = token;
-        Map<TruffleString, TruffleString> assertions = Map.of();
+        Map<TruffleString, TruffleString> attributes = Map.of();
         expect(EXPORT);
         final boolean yield = false;
         final boolean await = isTopLevelAwait();
@@ -7254,12 +7255,12 @@ public class Parser extends AbstractParser {
                     exportName = getIdentifierName();
                 }
                 FromNode from = fromClause();
-                if (env.importAssertions && type == ASSERT && last != EOL) {
-                    assertions = assertClause();
+                if (env.importAttributes && ((type == WITH) || (type == ASSERT && last != EOL))) {
+                    attributes = withClause();
                 }
                 TruffleString moduleRequest = from.getModuleSpecifier().getValue();
-                module.addModuleRequest(ModuleRequest.create(moduleRequest, assertions));
-                module.addExport(new ExportNode(exportToken, Token.descPosition(exportToken), finish, exportName, from, assertions));
+                module.addModuleRequest(ModuleRequest.create(moduleRequest, attributes));
+                module.addExport(new ExportNode(exportToken, Token.descPosition(exportToken), finish, exportName, from, attributes));
                 endOfLine();
                 break;
             }
@@ -7268,13 +7269,13 @@ public class Parser extends AbstractParser {
                 FromNode from = null;
                 if (type == FROM) {
                     from = fromClause();
-                    if (env.importAssertions && type == ASSERT && last != EOL) {
-                        assertions = assertClause();
+                    if (env.importAttributes && ((type == WITH) || (type == ASSERT && last != EOL))) {
+                        attributes = withClause();
                     }
                     TruffleString moduleRequest = from.getModuleSpecifier().getValue();
-                    module.addModuleRequest(ModuleRequest.create(moduleRequest, assertions));
+                    module.addModuleRequest(ModuleRequest.create(moduleRequest, attributes));
                 }
-                module.addExport(new ExportNode(exportToken, Token.descPosition(exportToken), finish, exportClause, from, assertions));
+                module.addExport(new ExportNode(exportToken, Token.descPosition(exportToken), finish, exportClause, from, attributes));
                 endOfLine();
                 break;
             }