+++ graal-js/src/com.oracle.truffle.js.parser/src/com/oracle/truffle/js/parser/GraalJSTranslator.java
@@ -2660,8 +2660,8 @@ abstract class GraalJSTranslator extends com.oracle.js.parser.ir.visitor.Transla
     }
 
     private JavaScriptNode createImportCallNode(JavaScriptNode[] args) {
-        assert args.length == 1 || (context.getLanguageOptions().importAssertions() && args.length == 2);
-        if (context.getLanguageOptions().importAssertions() && args.length == 2) {
+        assert args.length == 1 || (context.getLanguageOptions().importAttributes() && args.length == 2);
+        if (context.getLanguageOptions().importAttributes() && args.length == 2) {
             return factory.createImportCall(context, args[0], activeScriptOrModule, args[1]);
         }
         return factory.createImportCall(context, args[0], activeScriptOrModule);