+++ graal-js/src/com.oracle.truffle.js.test/js/module/import_call_assertion_error.mjs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
@@ -8,7 +8,7 @@
 /**
  * Tests that import calls with invalid options/assertions throw a TypeError.
  *
- * @option import-assertions=true
+ * @option import-attributes=true
  */
 
 function shouldHaveThrown(err) {